Add a custom 'usePlugDescriptions' hook that encapsulates the Clarity description fallback logic.
