+++ src/app/loadout/plug-drawer/SelectablePlug.tsx
@@ -3,8 +3,7 @@ import RichDestinyText from 'app/dim-ui/RichDestinyText';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { DefItemIcon } from 'app/inventory/ItemIcon';
 import { StatValue } from 'app/item-popup/PlugTooltip';
-import { useD2Definitions } from 'app/manifest/selectors';
-import { getPlugDescriptions } from 'app/utils/plug-descriptions';
+import { usePlugDescriptions } from 'app/utils/plug-descriptions';
 import clsx from 'clsx';
 import { useCallback, useMemo } from 'react';
 import styles from './SelectablePlug.m.scss';
@@ -68,8 +67,7 @@ function SelectablePlugDetails({
   plug: PluggableInventoryItemDefinition;
   displayedStatHashes?: number[];
 }) {
-  const defs = useD2Definitions()!;
-  const plugDescriptions = getPlugDescriptions(plug, defs);
+  const plugDescriptions = usePlugDescriptions(plug);
   const displayedStats = plug.investmentStats.filter((stat) =>
     displayedStatHashes?.includes(stat.statTypeHash)
   );
